core: add validation of relaxed path
